RawCosvFileController
### What's done:
- added a dedicated controller for raw cosv files

It closes #2713
