+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/basic/fileuploader/CosvFileManagerComponent.kt
@@ -75,7 +75,7 @@ val cosvFileManagerComponent: FC<Props> = FC { _ ->
     val deleteFile = useDeferredRequest {
         fileToDelete?.let { file ->
             val response = delete(
-                "$apiUrl/cosv/$selectedOrganization/delete/${file.requiredId()}",
+                "$apiUrl/raw-cosv/$selectedOrganization/delete/${file.requiredId()}",
                 jsonHeaders,
                 loadingHandler = ::loadingHandler,
             )
@@ -91,7 +91,7 @@ val cosvFileManagerComponent: FC<Props> = FC { _ ->
 
     val deleteProcessedFiles = useDeferredRequest {
         val response = delete(
-            "$apiUrl/cosv/$selectedOrganization/delete-processed",
+            "$apiUrl/raw-cosv/$selectedOrganization/delete-processed",
             jsonHeaders,
             loadingHandler = ::loadingHandler,
         )
@@ -122,7 +122,7 @@ val cosvFileManagerComponent: FC<Props> = FC { _ ->
         selectedOrganization?.let {
             val newPage = lastPage.inc()
             val response = get(
-                url = "$apiUrl/cosv/$selectedOrganization/list",
+                url = "$apiUrl/raw-cosv/$selectedOrganization/list",
                 params = jso<dynamic> {
                     page = newPage - 1
                     size = DEFAULT_SIZE
@@ -153,7 +153,7 @@ val cosvFileManagerComponent: FC<Props> = FC { _ ->
     val reFetchFiles = useDeferredRequest {
         selectedOrganization?.let {
             val count: Long = get(
-                url = "$apiUrl/cosv/$selectedOrganization/count",
+                url = "$apiUrl/raw-cosv/$selectedOrganization/count",
                 jsonHeaders,
                 loadingHandler = ::loadingHandler,
                 responseHandler = ::noopResponseHandler
@@ -168,7 +168,7 @@ val cosvFileManagerComponent: FC<Props> = FC { _ ->
     val uploadFiles = useDeferredRequest {
         setStreamingOperationActive(true)
         val response = post(
-            url = "$apiUrl/cosv/$selectedOrganization/batch-upload",
+            url = "$apiUrl/raw-cosv/$selectedOrganization/batch-upload",
             headers = Headers().withAcceptNdjson(),
             body = FormData().apply { filesForUploading.forEach { append(FILE_PART_NAME, it) } },
             loadingHandler = ::noopLoadingHandler,
@@ -194,7 +194,7 @@ val cosvFileManagerComponent: FC<Props> = FC { _ ->
         fileToUnzip?.let { file ->
             setStreamingOperationActive(true)
             val response = post(
-                "$apiUrl/cosv/$selectedOrganization/unzip/${file.requiredId()}",
+                "$apiUrl/raw-cosv/$selectedOrganization/unzip/${file.requiredId()}",
                 headers = Headers().withContentTypeJson().withAcceptNdjson(),
                 body = undefined,
                 loadingHandler = ::noopLoadingHandler,
@@ -233,7 +233,7 @@ val cosvFileManagerComponent: FC<Props> = FC { _ ->
 
     val submitCosvFiles = useDeferredRequest {
         val response = post(
-            url = "$apiUrl/cosv/$selectedOrganization/submit-to-process",
+            url = "$apiUrl/raw-cosv/$selectedOrganization/submit-to-process",
             jsonHeaders,
             body = selectedFiles.map { it.requiredId() },
             loadingHandler = ::loadingHandler,
@@ -247,7 +247,7 @@ val cosvFileManagerComponent: FC<Props> = FC { _ ->
 
     val submitAllUploadedCosvFiles = useDeferredRequest {
         val response = post(
-            url = "$apiUrl/cosv/$selectedOrganization/submit-all-uploaded-to-process",
+            url = "$apiUrl/raw-cosv/$selectedOrganization/submit-all-uploaded-to-process",
             jsonHeaders,
             body = undefined,
             loadingHandler = ::loadingHandler,
@@ -348,7 +348,7 @@ val cosvFileManagerComponent: FC<Props> = FC { _ ->
                         }
                     }
                     downloadFileButton(file, RawCosvFileDto::fileName) {
-                        "$apiUrl/cosv/$selectedOrganization/download/${file.requiredId()}"
+                        "$apiUrl/raw-cosv/$selectedOrganization/download/${file.requiredId()}"
                     }
                     if (file.fileName.endsWith(ARCHIVE_EXTENSION, ignoreCase = true)) {
                         button {
